requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
ca-certificates=2019.3.9=hecc5488_0
certifi=2019.3.9=py36_0
cffi=1.12.2=py36hf0e25f4_1
cuda75=1.0=hf2493ae_0
cycler=0.10.0=py_1
dbus=1.13.2=h714fa37_1
expat=2.2.5=hf484d3e_1002
fontconfig=2.13.1=he4413a7_1000
freetype=2.10.0=he983fc9_0
gettext=0.19.8.1=hc5be6a0_1002
glib=2.56.2=had28632_1001
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
icu=58.2=hf484d3e_1000
intel-openmp=2019.3=199
joblib=0.13.2=pypi_0
jpeg=9c=h14c3975_1001
kiwisolver=1.0.1=py36h6bb024c_1002
libblas=3.8.0=4_openblas
libcblas=3.8.0=4_openblas
libffi=3.2.1=he1b5a44_1006
libgcc-ng=8.2.0=hdf63c60_1
libgfortran=3.0.0=1
libiconv=1.15=h516909a_1005
libpng=1.6.36=h84994c4_1000
libprotobuf=3.7.1=h8b12597_0
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.0.10=h648cc4a_1001
libuuid=2.32.1=h14c3975_1000
libxcb=1.13=h14c3975_1002
libxml2=2.9.8=h143f9aa_1005
matplotlib=3.0.3=py36_0
matplotlib-base=3.0.3=py36h167e16e_0
mkl=2019.3=199
ncurses=6.1=hf484d3e_1002
ninja=1.9.0=h6bb024c_0
numpy=1.16.2=py36h7fc5cc3_0
olefile=0.46=py_0
openblas=0.3.5=ha44fe06_0
openssl=1.1.1b=h14c3975_1
pcre=8.43=he6710b0_0
pillow=6.0.0=py36he7afcd5_0
pip=19.0.3=py36_0
protobuf=3.7.1=py36he1b5a44_0
pthread-stubs=0.4=h14c3975_1001
pycparser=2.19=py36_1
pyparsing=2.4.0=py_0
pyqt=5.6.0=py36h13b7fb3_1008
python=3.6.7=h381d211_1004
python-dateutil=2.8.0=py_0
pytorch=0.4.1=py36_py35_py27__9.0.176_7.1.2_2
qt=5.6.3=h8bf5577_3
readline=7.0=hf8c457e_1001
scikit-learn=0.20.3=pypi_0
scikit-plot=0.3.7=pypi_0
scipy=1.2.1=pypi_0
setuptools=41.0.0=py36_0
sip=4.18.1=py36hf484d3e_1000
six=1.12.0=py36_1000
sqlite=3.26.0=h67949de_1001
tensorboardx=1.6=py_0
tk=8.6.9=h84994c4_1001
torchvision=0.2.1=py_2
tornado=6.0.2=py36h516909a_0
wheel=0.33.1=py36_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xz=5.2.4=h14c3975_1001
zlib=1.2.11=h14c3975_1004
